varios detalles

+++ application/controllers/Producto.php
@@ -295,7 +295,9 @@ class Producto extends CI_Controller{
                           unlink($directorio.$foto1);
                           //$mimagenthumb = str_replace(".", "_thumb.", $foto1);
                           $mimagenthumb = "thumb_".$foto1;
-                          unlink($directorio.$mimagenthumb);
+                          if(file_exists($mimagenthumb)){
+                              unlink($directorio.$mimagenthumb);
+                          }
                       }
                   }
                     $confi['image_library'] = 'gd2';
